refactor: create UserAccount component
